Revert "drivers: ethernet: eth_stm32 v1/v2 are placed in separate files"
This reverts commit 29750ac05de52ffe0dc2873c5de4f261001f3b31.
This commit was part of PR 94142 which has been merged
while a discussion was still on going on the right direction
between splitting V1 and V2 or keeping a common file.
