+++ drivers/ethernet/CMakeLists.txt
@@ -29,8 +29,7 @@ zephyr_library_sources_ifdef(CONFIG_ETH_ESP32		eth_esp32.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_LITEX_LITEETH	eth_litex_liteeth.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SMSC911X	eth_smsc911x.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_STELLARIS	eth_stellaris.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL_API_V1	eth_stm32_hal_v1.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL_API_V2	eth_stm32_hal_v2.c)
+zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL	eth_stm32_hal_v2.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_W5500		eth_w5500.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SAM_GMAC	eth_sam_gmac.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_CYCLONEV	eth_cyclonev.c)
